clang-tidy misc-definitions-in-headers (#33170)
* Move wcwidth implementation to cpp file

We had the whole wcwidth implementation in wcwidth.h, including
functions that were never used.  These are sufficiently complex
functions that inlining for performance reasons serves little purpose.

Move them to a new wcwidth.cpp file, and delete the unused functions.

* Enable misc-definitions-in-headers

* Don't declare mk_wcwidth in catacharset.h

To keep things organized it should only be declared in the header
corresponding to its cpp file (wcwidth.h).
