+++ src/catacharset.h
@@ -21,8 +21,6 @@ inline uint32_t UTF8_getch( const std::string &str )
     int len = str.length();
     return UTF8_getch( &utf8str, &len );
 }
-// from wcwidth.c, return "cell" width of a Unicode char
-int mk_wcwidth( uint32_t ucs );
 // convert cursorx value to byte position
 int cursorx_to_position( const char *line, int cursorx, int *prevpos = nullptr, int maxlen = -1 );
 int utf8_width( const char *s, const bool ignore_tags = false );
